Fixed Aggregate scaling problem
We do not want individual nodes to take into account the usage of other nodes.
Therefore, we send the distribution of the basis of U to the
aggregator. We then map the current resource usage matrix to the
aggregate basis and calculate the sum of the squares of the components
to get the "singular values".